+++ lib/src/chatruimte.dart
@@ -1,12 +1,12 @@
 import 'package:flutter/material.dart';
 import 'authentication.dart';
-import 'guestbook.dart';
 import 'widgets.dart';
 import 'application.dart';
 import 'package:provider/provider.dart';
+import 'dart:async';
 
 class ChatRuimte extends StatelessWidget {
-//  HomePage({Key? key}) : super(key: key);
+  ChatRuimte({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -16,10 +16,6 @@ class ChatRuimte extends StatelessWidget {
       ),
       body: ListView(
         children: <Widget>[
-//          Header("Gebruik de nieuwe App"),
-//          Paragraph(
-//            'Twee dagen strijden alle klassen tegen elkaar',
-//          ),
           Consumer<ApplicationState>(
             builder: (context, appState, _) => Column(
               crossAxisAlignment: CrossAxisAlignment.start,
@@ -37,6 +33,79 @@ class ChatRuimte extends StatelessWidget {
           ),
         ],
       ),
+    );
+  }
+}
+
+class GuestBookMessage {
+  GuestBookMessage({required this.name, required this.message});
+  final String name;
+  final String message;
+}
+
+class GuestBook extends StatefulWidget {
+  GuestBook({required this.addMessage, required this.messages});
+  final FutureOr<void> Function(String message) addMessage;
+  final List<GuestBookMessage> messages;
+
+  @override
+  _GuestBookState createState() => _GuestBookState();
+}
+
+class _GuestBookState extends State<GuestBook> {
+  final _formKey = GlobalKey<FormState>(debugLabel: '_GuestBookState');
+  final _controller = TextEditingController();
+
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: [
+        Padding(
+          padding: const EdgeInsets.all(8.0),
+          child: Form(
+            key: _formKey,
+            child: Row(
+              children: [
+                Expanded(
+                  child: TextFormField(
+                    controller: _controller,
+                    decoration: const InputDecoration(
+                      hintText: 'typ je bericht',
+                    ),
+                    validator: (value) {
+                      if (value == null || value.isEmpty) {
+                        return 'Voer een bericht in om door te gaan';
+                      }
+                      return null;
+                    },
+                  ),
+                ),
+                SizedBox(width: 8),
+                StyledButton(
+                  onPressed: () async {
+                    if (_formKey.currentState!.validate()) {
+                      await widget.addMessage(_controller.text);
+                      _controller.clear();
+                    }
+                  },
+                  child: Row(
+                    children: [
+                      Icon(Icons.send),
+                      SizedBox(width: 4),
+                      Text('verzend'),
+                    ],
+                  ),
+                ),
+              ],
+            ),
+          ),
+        ),
+        SizedBox(height: 8),
+        for (var message in widget.messages)
+          Paragraph('${message.name}: ${message.message}'),
+        SizedBox(height: 8),
+      ],
     );
   }
 }
